Tests/LibWeb/Text/expected/abspos-relayout-basic.txt

child=40,30,70,60
outside-before=0,120,200,25
outside-after=0,120,200,25
partial-layouts=0 full-layouts=1
FAIL
